windows fixups 3

+++ apps/thirdeye/graphics/palette.cpp
@@ -22,7 +22,7 @@ GRAPHICS::Palette::Palette(const std::vector<uint8_t> &pal, bool isRes) {
 	} else {
 		mNumOfColours = pal.size()/3; // 3 is rgb
 		uint16_t counter = 0;
-		for(uint i=0; i<pal.size(); i+=3){
+		for(uint16_t i=0; i<pal.size(); i+=3){
 			// Bitshift from 8 bits to 6 bits that is which is our palette size
 			mPalette[counter].r = pal[i] << 2;
 			mPalette[counter].g = pal[i+1] << 2;
